requirements.txt
discord.py
python-dotenv
pandas
openpyxl
mysql-connector-python
langchain
langgraph
langchain-google-genai
google-genai
PyPDF2